requirements.txt
streamlit == 1.50.0
torch == 2.8.0
torchvision==0.23.0
pillow==11.3.0